backend/data-preprocessing-service/requirements.txt
aioredis==2.0.1
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
joblib==1.4.2
numpy==2.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
